requirements.txt
requests>=2.28.1
beautifulsoup4>=4.11.1
fake-useragent>=1.1.3
python-dotenv>=0.21.0
selenium>=4.8.0
webdriver-manager>=3.8.5